合同卸货费

+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/pojo/Contract.java
@@ -42,7 +42,7 @@ public class Contract extends CommonBean {
     private Byte deliveryCharges;
 
     @ApiModelProperty("卸货费(甲方，乙方承担)")
-    private BigDecimal unloadingFee;
+    private Long unloadingFee;
 
     @ApiModelProperty("其他约定")
     private String otherConventions;
