+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/request/applycontract/vo/ApplyContractReqVo.java
@@ -59,7 +59,7 @@ public class ApplyContractReqVo{
 
     @ApiModelProperty("卸货费(甲方，乙方承担)")
     @NotNull(message = "卸货费不能为空")
-    private BigDecimal unloadingFee;
+    private Long unloadingFee;
 
     @ApiModelProperty("其他约定")
     private String otherConventions;
