+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/request/contract/dto/ContractDTO.java
@@ -50,7 +50,7 @@ public class ContractDTO extends CommonBean {
     private Byte deliveryCharges;
 
     @ApiModelProperty("卸货费(甲方，乙方承担)")
-    private BigDecimal unloadingFee;
+    private Long unloadingFee;
 
     @ApiModelProperty("其他约定")
     private String otherConventions;
