+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/request/contract/vo/ContractReqVo.java
@@ -61,7 +61,7 @@ public class ContractReqVo {
 
     @ApiModelProperty("卸货费(甲方，乙方承担)")
     @NotNull(message = "卸货费承担方不能为空")
-    private BigDecimal unloadingFee;
+    private Long unloadingFee;
 
     @ApiModelProperty("其他约定")
     private String otherConventions;
