+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/response/applycontract/ApplyContractViewResVo.java
@@ -57,7 +57,7 @@ public class ApplyContractViewResVo {
     private Byte deliveryCharges;
 
     @ApiModelProperty("卸货费(甲方，乙方承担)")
-    private BigDecimal unloadingFee;
+    private Long unloadingFee;
 
     @ApiModelProperty("其他约定")
     private String otherConventions;
